review: finish remaining hardening fixes
Align protocol tab indices with the FlatBuffers u32 contract across the CLI, client, server, and snapshot decoding paths, and keep PTY helper threads owned by the buffer runtime so shutdown and exit observe thread failures instead of dropping them silently.

+++ crates/embers-client/tests/reducer.rs
@@ -47,7 +47,7 @@ fn buffer_view_node(
     }
 }
 
-fn session_snapshot(root_active: usize, nested_active: usize) -> SessionSnapshot {
+fn session_snapshot(root_active: u32, nested_active: u32) -> SessionSnapshot {
     let session_id = SessionId(1);
     SessionSnapshot {
         session: SessionRecord {
